css/css-scroll-snap/scroll-snap-type-proximity.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
div {
  position: absolute;
  margin: 0px;
}
#scroller {
  height: 600px;
  width: 600px;
  overflow: scroll;
}
#space {
  width: 2000px;
  height: 2000px;
}

#snap {
  width: 200px;
  height: 200px;
  left: 800px;
  top: 800px;
  background-color: blue;
  scroll-snap-align: start;
}
</style>

<div id="scroller">
  <div id="space"></div>
  <div id="snap"></div>
</div>

<script>
var scroller = document.getElementById("scroller");
var visible_x = 800 - scroller.clientWidth;
var visible_y = 800 - scroller.clientHeight;

test(() => {
  scroller.style.scrollSnapType = "both mandatory";
  scroller.scrollTo(visible_x + 10, visible_y + 10);
  assert_equals(scroller.scrollLeft, 800);
  assert_equals(scroller.scrollTop, 800);
}, "mandatory scroll-snap-type should snap as long as the element is visible.");

test(() => {
  scroller.style.scrollSnapType = "both proximity";
  scroller.scrollTo(visible_x + 10, visible_y + 10);
  assert_equals(scroller.scrollLeft, visible_x + 10);
  assert_equals(scroller.scrollTop, visible_y + 10);
}, "proximity scroll-snap-type shouldn't snap if the snap position is too far away.");

test(() => {
  scroller.style.scrollSnapType = "both proximity";
  scroller.scrollTo(795, 795);
  assert_equals(scroller.scrollLeft, 800);
  assert_equals(scroller.scrollTop, 800);
}, "proximity scroll-snap-type should snap if the snap position is close.");
</script>
